snapshot: rename volmgr_service.py to service.py for volmgr plugin
